pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "perfume_trend_sdk"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml",
    "python-dotenv",
    "pytest",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "httpx>=0.24",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
]

[tool.setuptools.packages.find]
where = ["."]
